Add support for resuming an upload that failed with a retryable error
This is achieved by keeping track of the status of the upload through a
state-file, which represents the complete state required for tracking
the status of an upload.

If the upload fails with a retryable error, the state-file will be
written out and is then usable through the `resume` command. If resume
makes its way further along, but still fails with (another) retryable
error, it can be invoked again, hopefully until the upload is fully
done.

In case there is an unrecoverable error, the multi-part upload will be
aborted.