Ensure `State` is not cloneable
If we ever implement parallel uploads, it is important that there is
only one state throughout the application, and only one state that is
ever persisted to disk. By making `State` not cloneable, we require that
modification to the state is done through some kind of primitive that
ensures mutual exlusivity.

(Further modifications to the `State` are of course required if we
introduce parallelism, fields like `last_successful_part` for example
wouldn't make sense anymore. Still, we introduce this change now to make
a transition to supporting parallel uploads potentially a little
easier.)