Write state-file whenever a part was (attempted to be) uploaded
If we only write the state when something (retryable) fails, we do not
support resuming an upload if the application itself crashed, or the
system it was running on.

We now persist the state whenever a part was (attempted to be) uploaded,
which ensures that even after a system crash we can likely resume an
upload instead of having to start it from scratch.